+++ myteststuff/simple.cpp
@@ -1,0 +1,8 @@
+//
+// Created by dfekete on 2024.03.15..
+//
+
+void hello_cppp()
+{
+  int i = 1 + 2;
+}