+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -161,6 +161,21 @@ New checks
 - New :doc:`bugprone-assignment-in-if-condition
   <clang-tidy/checks/bugprone/assignment-in-if-condition>` check.
 
+- New :doc:`modernize-use-ranges
+  <clang-tidy/checks/modernize/use-ranges>` check.
+
+  FIXME: add release notes.
+
+- New :doc:`readability-early-exit
+  <clang-tidy/checks/readability/early-exit>` check.
+
+  FIXME: add release notes.
+
+- New :doc:`readability-simplify-bool-expressions
+  <clang-tidy/checks/readability/simplify-bool-expressions>` check.
+
+  FIXME: add release notes.
+
   Warns when there is an assignment within an if statement condition expression.
 
 - New :doc:`misc-const-correctness